+++ macros/CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add methods `client()`, `id()`, `version()`, `timestamp()`, `service()`, `abort()`,
   `is_aborted()`, `poll_aborted()` and `aborted()` to generated call enums.
 - Add methods `id()`, `timestamp()` and `service()` to generated event enums.
+- There is now special support for newtypes via the attribute `#[aldrin(newtype)]`, which can be
+  applied to structs with a single field. Such types will then serialize directly as their inner
+  field.
 
 ### Changed